requirements.txt
django
flake8
flake8-quotes
flake8-variables-names
pep8-naming
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
drf-spectacular
Pillow
psycopg
psycopg-binary
python-dotenv